fix 0.4.0 compilation errors.

+++ react-streams-vertx/src/main/java/com/github/bckfnn/reactstreams/vertx/RsEventBus.java
@@ -25,7 +25,7 @@ public class RsEventBus {
 			subscriber.onSubscribe(new BaseSubscription<JsonObject>(subscriber) {
 				boolean done = false;
 				@Override
-				public void request(int elements) {
+				public void request(long elements) {
 					System.out.println("request:" + elements);
 					super.request(elements);
 					if (done) {
